Design/fe/kdh

See merge request s08-ai-speech-sub2/S08P22A703!140

+++ frontend/src/components/pronounce/FreqUsedPhrase.vue
@@ -2,17 +2,24 @@
   <div class="TTS__container">
     <div class="cat__wrapper">
       <div class="cat__container">
-        <div class="cat__item" v-for="(category, c) in freqUsedCat" :key="c">
-          <p
-            style="width: max-content"
-            @click="phraseHandler(category.categoryId)"
-          >
+        <div
+          class="cat__item"
+          :id="`${index}`"
+          v-for="(category, index) in freqUsedCat"
+          :key="index"
+          @click="
+            (e) => {
+              activeHandler(index), phraseHandler(category.categoryId);
+            }
+          "
+        >
+          <p>
             {{ category.categoryName }}
           </p>
         </div>
       </div>
       <div class="catadd__button">
-        <font-awesome-icon @click="addCatModal" icon="fa-solid fa-gear" />
+        <font-awesome-icon @click="addCatModal" icon="fa-solid fa-plus" />
       </div>
     </div>
     <AddCatModal v-if="isCatModal" @closemodal="addCatModal" />
@@ -23,7 +30,13 @@
         :key="index"
       >
         <div>
-          <p @click="getAudio(phrase.sentence)">{{ phrase.sentence }}</p>
+          <p
+            :id="`p${index}`"
+            class="p_class"
+            @click="getAudio(phrase.sentence, index)"
+          >
+            {{ phrase.sentence }}
+          </p>
         </div>
         <div>
           <font-awesome-icon
@@ -45,45 +58,39 @@
 </template>
 
 <script setup>
-import { ref, watch } from "vue";
+import { ref } from "vue";
 import AddCatModal from "./AddCatModal.vue";
 import AddPhraseModal from "./AddPhraseModal.vue";
 import { usePronounce } from "@/store/Pronounce.js";
 import { storeToRefs } from "pinia";
+import Swal from "sweetalert2";
 
 // data from store
 const store = usePronounce();
 const { freqUsedCat, freqUsedPhrase } = storeToRefs(store);
 
-// 페이지 타입 구분할 변수
-const pageIndex = 1;
+const idIndex = ref(undefined);
+const parseCategoryId = ref(undefined);
+const isCatModal = ref(false);
+const isPhraseModal = ref(false);
+const audio = ref();
 
 // api 호출 함수 실행 시키는 함수
 const callCategoryAPI = async () => {
   await store.readCategories;
+  if (freqUsedCat.value.length) {
+    idIndex.value = 0;
+    document.getElementById(idIndex.value).className = "cat__item touched";
+    phraseHandler(freqUsedCat.value[0].categoryId);
+  }
 };
 const callSentenceAPI = async () => {
-  console.log("callsentence API", parseCategoryId.value);
   await store.readQuickSentence(parseCategoryId.value);
 };
-
-// 스토어 값이 변할 때마다 실행되어야함,
-watch(freqUsedPhrase, () => {
-  // callSentenceAPI();
-  console.log("watch sentence", freqUsedPhrase.value);
-});
-watch(freqUsedCat, () => {
-  // callAPI();
-  // console.log("watch category", freqUsedCat);
-});
-
-const isCatModal = ref(false);
-const isPhraseModal = ref(false);
 
 // catagory add modal
 const addCatModal = () => {
   callCategoryAPI();
-  // store.readQuickSentence();
 
   if (isPhraseModal.value == true) {
     isPhraseModal.value = false;
@@ -102,36 +109,53 @@ const addPhraseModal = () => {
 };
 
 // category 별 phrase 불러오는 함수 + categoryID 저장하기
-const parseCategoryId = ref(1);
+
 const phraseHandler = (categoryId) => {
-  console.log(categoryId);
+  // console.log(categoryId);
   store.readQuickSentence(categoryId);
   parseCategoryId.value = categoryId;
 };
 
 // Delete 자주쓰는문장 Handler
 const delPhraseHandler = (sentenceId) => {
+  Swal.fire({
+    title: "삭제 하시겠습니까?",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      store.deleteQuickSentence(sentenceId);
+    }
+  });
   // alert("삭제 되었습니다.");
-  store.deleteQuickSentence(sentenceId);
   callSentenceAPI();
 };
 
 // 선택한 목소리로 TTS 실행하기
-const getAudio = (text) => {
-  console.log("TTS test", text);
-  let audio = new Audio(
-    `http://j8a703.p.ssafy.io/ai/infer/?text=${encodeURI(text)}`
+const getAudio = (text, sentenceIndex) => {
+  document.getElementById(`p${sentenceIndex}`).className = "p_touched";
+  audio.value = new Audio(
+    `http://j8a703.p.ssafy.io/ai/infer/?text=${encodeURI(text)}&voice=1`
   );
-  audio.play();
+  audio.value.play();
+  audio.value.onended = function () {
+    document.getElementById(`p${sentenceIndex}`).className = "p_class";
+  };
 };
-// const playTTS = () => {
-//   //AI 함수 호출는 부분
-//   console.log("AI TTS 실행시키기");
-//   getAudio();
-// };
+
+const activeHandler = (index) => {
+  if (idIndex.value !== undefined) {
+    document.getElementById(idIndex.value).className = "cat__item";
+  }
+  document.getElementById(index).className = "cat__item touched";
+  idIndex.value = document.getElementById(index).getAttribute("id");
+};
+
 // GET catagoires
 callCategoryAPI();
-callSentenceAPI();
 </script>
 
 <style lang="scss" scoped>
@@ -152,10 +176,18 @@ callSentenceAPI();
   overflow-x: scroll;
 }
 .cat__item {
+  // border-top: 5px solid var(--maincolor2);
   display: flex;
-  width: 50px;
-  padding-left: 6%;
-  height: 100%;
+  width: fit-content;
+  margin-top: -3%;
+  padding-left: 3%;
+  padding-right: 3%;
+  height: 70%;
+  border-radius: 5px;
+  align-items: center;
+}
+p {
+  text-align: center;
 }
 .catadd__button {
   padding-right: 4%;
@@ -163,12 +195,12 @@ callSentenceAPI();
   justify-self: center;
 }
 .TTS__container {
+  border: solid var(--maincolor2);
   background-color: var(--black1);
   margin-inline: 2vw;
   margin-bottom: 2vh;
   padding: 1vh;
   border-radius: 16px;
-  border: solid var(--maincolor2) 3px;
 }
 .phrase__box {
   padding-inline: 3vw;
@@ -180,5 +212,11 @@ callSentenceAPI();
 }
 .freqButton__box {
   padding-inline: 3vw;
+}
+.touched {
+  background-color: var(--maincolor2);
+}
+.p_touched {
+  color: var(--maincolor4);
 }
 </style>
